dom/virtualfilesystem/test/mochitest/test_virtualfilesystem_event_remote.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1026350
-->
<head>
<meta charset="utf-8">
<title>Test Inputport Connection Event</title>
<link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1187723">Test FileSystemProvider Event
</a>
<script type="application/javascript;version=1.8">

"use strict";

SimpleTest.waitForExplicitFinish();
function MockCallback() {
}

function testEvent() {
  MockCallback.prototype = {
      onSuccess: function(requestId, value, hasMore) {},
      onError: function(requestId, error) {}
  };
  return new Promise(function(resolve, reject) {
    var Ci = SpecialPowers.Ci;

    var fileSystemProvider = navigator.fileSystemProvider;
    var _requestId = 1;
    var _operationRequestId = 1;
    fileSystemProvider.addEventListener("abortrequested", function onabort(event){
      fileSystemProvider.removeEventListener("abortrequested", onabort);
      ok(true, "got event.");
      is(event.options.fileSystemId, _fileSystemId, "file system id must be the same");
      is(event.options.requestId, _requestId, "request id must be the same");
      is(event.options.operationRequestId, _operationRequestId, "operation request id must be the same");
      fileSystemProvider.unmount({fileSystemId: _fileSystemId}).then(function() {
          resolve();
        }
      );
    });
    fileSystemProvider.mount({fileSystemId: _fileSystemId,
                              displayName: _displayName,
                              writable: _writable,
                              openedFilesLimit: _openedFilesLimit}).then(
      function() {
        ok(true, "mount ok.");
        var vfsService;
        if (SpecialPowers.isMainProcess()) {
          vfsService = SpecialPowers.Cc["@mozilla.org/virtualfilesystem/fakevirtualfilesystemservice;1"].getService(Ci.nsIVirtualFileSystemService);
        }
        else {
          vfsService = SpecialPowers.Cc["@mozilla.org/virtualfilesystem/virtualfilesystem-service;1"].getService(Ci.nsIVirtualFileSystemService);
        }
        ok(vfsService, "must have vfsService");
        var requestManager = vfsService.getRequestManagerById(_fileSystemId);
        ok(requestManager, "must have request manager.");

        var abortOptions = SpecialPowers.Cc["@mozilla.org/virtualfilesystem/virtualfilesystem-abort-requested-options;1"].createInstance(Ci.nsIVirtualFileSystemAbortRequestedOptions);
        abortOptions.fileSystemId = _fileSystemId;
        abortOptions.requestId = _requestId;
        abortOptions.operationRequestId = _operationRequestId;
        requestManager.createRequest(Ci.nsIVirtualFileSystemRequestManager.REQUEST_ABORT,
                                     abortOptions,
                                     new MockCallback());
      },
      function(aError) {
        ok(false, "Fail to mount: " + aError);
        resolve();
      }
    );
  });
}


var iframeUrl = SimpleTest.getTestFileURL('file_virtualfilesystem_oop.html');

function runTest() {
  SpecialPowers.addPermission("filesystemprovider",
                              true, { url: iframeUrl,
                                      originAttributes: {
                                        appId: SpecialPowers.Ci.nsIScriptSecurityManager.NO_APP_ID,
                                        inBrowser: true }});
  var iframe = document.createElement("iframe");
  iframe.setAttribute("mozbrowser", "true");
  iframe.setAttribute("remote", "true");
  iframe.setAttribute("src", iframeUrl);

  iframe.addEventListener('mozbrowsershowmodalprompt', function receiverListener(aEvent) {
    var message = aEvent.detail.message;
    if (/^OK /.exec(message)) {
      ok(true, "Message from iframe: " + message);
    } else if (/^KO /.exec(message)) {
      ok(false, "Message from iframe: " + message);
    } else if (/^INFO /.exec(message)) {
      info("Message from iframe: " + message);
    } else if (/^COMMAND /.exec(message)) {
    } else if (/^DONE$/.exec(message)) {
      ok(true, "Messaging from iframe complete.");
      iframe.removeEventListener('mozbrowsershowmodalprompt', receiverListener);
      SimpleTest.finish();
    }
  }, false);

  document.body.appendChild(iframe);
}

SpecialPowers.pushPrefEnv({"set": [["device.filesystemprovider.enabled", true],
                                    //to ignore app scope check.
                                   ["dom.ignore_webidl_scope_checks", true],
                                   ["dom.mozBrowserFramesEnabled", true],
                                   ["dom.ipc.tabs.disabled", false]]}, function() {
  SpecialPowers.pushPermissions(
    [{type: "browser", allow: true, context: document },
     {type: 'filesystemprovider', allow: true, context: document},], runTest);
});

 </script>
 </pre>
 </body>
 </html>
